+++ packages/auth0-server-js/src/state/utils.ts
@@ -2,6 +2,19 @@ import type { AccessTokenForConnectionOptions, StateData } from '../types.js';
 import { TokenResponse } from '@auth0/auth0-auth-js';
 
 /**
+ * Creates an updated token set object from the token endpoint response
+ * @param audience The audience for which the token was requested
+ * @param response The response from the token endpoint
+ * @returns Updated token set object
+ */
+const createUpdatedTokenSet = (audience: string, response: TokenResponse) => ({
+  audience,
+  accessToken: response.accessToken,
+  scope: response.scope,
+  expiresAt: response.expiresAt,
+});
+
+/**
  * Utility function to update the state with a new response from the token endpoint
  * @param audience The audience of the token endpoint response
  * @param stateDate The existing state data to update, or undefined if no state data available.
@@ -18,18 +31,11 @@ export function updateStateData(
       (tokenSet) => tokenSet.audience === audience && tokenSet.scope === tokenEndpointResponse.scope
     );
 
-    const createUpdatedTokenSet = (response: TokenResponse) => ({
-      audience,
-      accessToken: response.accessToken,
-      scope: response.scope,
-      expiresAt: Math.floor(Date.now() / 1000) + Number(response.expiresAt),
-    });
-
     const tokenSets = isNewTokenSet
-      ? [...stateDate.tokenSets, createUpdatedTokenSet(tokenEndpointResponse)]
+      ? [...stateDate.tokenSets, createUpdatedTokenSet(audience, tokenEndpointResponse)]
       : stateDate.tokenSets.map((tokenSet) =>
           tokenSet.audience === audience && tokenSet.scope === tokenEndpointResponse.scope
-            ? createUpdatedTokenSet(tokenEndpointResponse)
+            ? createUpdatedTokenSet(audience, tokenEndpointResponse)
             : tokenSet
         );
 
@@ -46,12 +52,7 @@ export function updateStateData(
       idToken: tokenEndpointResponse.idToken,
       refreshToken: tokenEndpointResponse.refreshToken,
       tokenSets: [
-        {
-          audience,
-          accessToken: tokenEndpointResponse.accessToken,
-          scope: tokenEndpointResponse.scope,
-          expiresAt:tokenEndpointResponse.expiresAt,
-        },
+        createUpdatedTokenSet(audience, tokenEndpointResponse),
       ],
       internal: {
         sid: user?.sid as string,
@@ -84,8 +85,7 @@ export function updateStateDataForConnectionTokenSet(
   const connectionTokenSets = isNewTokenSet
     ? [...stateDate.connectionTokenSets, connectionTokenSet]
     : stateDate.connectionTokenSets.map((tokenSet) =>
-        tokenSet.connection === options.connection &&
-        (!options.loginHint || tokenSet.loginHint === options.loginHint)
+        tokenSet.connection === options.connection && (!options.loginHint || tokenSet.loginHint === options.loginHint)
           ? connectionTokenSet
           : tokenSet
       );